Move WorldRenderScreen into include/, src/ files
