+++ main.cpp
@@ -8,7 +8,7 @@
 #include <target_id.hpp>
 #include <inventory_gui/screen.hpp>
 #include <start_screen_gui/screen.hpp>
-#include "world_render.hpp"
+#include <world_render_screen.hpp>
 #include <world_navigation_gui/screen.hpp>
 #include <script.hpp>
